+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ✅ Added
 
+- Added haptic and system sound feedback for voice recording interactions. Feedback can be
+  customized by extending `AudioRecorderFeedback` or using `AudioRecorderFeedbackWrapper` with
+  custom callbacks. [[#2463]](https://github.com/GetStream/stream-chat-flutter/issues/2463)
 - Added support for deleting options while creating a poll in
   `StreamPollCreator`. [[#2453]](https://github.com/GetStream/stream-chat-flutter/issues/2453)